Add README documentation for JavaSorts program
A README.md file has been added to provide comprehensive information about the JavaSorts program. This file explains the purpose of the program, the core classes involved and describes the available sorting algorithms. Instructions on how to use the program are also provided. This documentation is necessary for other developers to understand the project structure and quickly get started.