Add logging and enhance test documentation
Logger was added to TimeUtilsTest, SortingAlgorithmsTest and SortingAlgorithmTest classes to provide more detailed information about testing process. Also, Javadoc comments were added to describe the purpose of test classes and methods thus improving readability. These changes are intended to improve the clarity of tests and increase traceability when debugging or enhancing the code.