"Added Javadoc comments and updated method ordering
Javadoc comments were added to `TimeUtils` class for educating developers about the utility of this class. A new parameter `outputSb` was added to the `displayAndTime` method in `SortingAlgorithm` to allow output to be printed to console or appended to a StringBuilder object based on need. Unnecessary comment lines in `HeapSort` were removed to streamline the code. All modifications have been made for better readability and self-documentation for future developers."