"Added empty list check in RadixSort algorithm
In RadixSort algorithm, added condition check for empty list. If the list is empty, -1 will be returned instead of trying to find the maximum value. Earlier, it was trying to sort the empty list which was not efficient. This update improves performance for the cases where the list is empty."