"Added TimSort to sorting algorithms
A new Java file, TimSort.java was added under the practicalsorts package to include the implementation of the TimSort algorithm. TimSort was also added to the list of Sorting algorithms in JavaSortsMain.java, SortingAlgorithmsTest.java, and SortingAlgorithmFactory.java. With this update, users can now select TimSort from the provided list of sorting algorithms, thereby expanding their selection options with a more efficient and practical sorting algorithm for large data sets."