"Refactor test and utility classes, Improve readability in sort classes
Reduced visibility to package-private for test methods and classes to promote encapsulation, as these do not necessarily need to be exposed as public. Added a private constructor in TimeUtils class to prevent its instantiation since the class only contains static methods.

The statements grouping in the merge methods of TimSort and MergeSort classes have been restructured for better readability and understanding. An additional assertion in SortingAlgorithmTest was included to verify that an empty list is returned when the requested size is 0."